ACAD_PROXY_ENTITY: more corrections to member types in dxf_acad_proxy_entity_struct.

+++ src/acad_proxy_entity.h
@@ -153,7 +153,7 @@ dxf_acad_proxy_entity_struct
                 /*!< Hard pointer ID / handle of PlotStyleName object.\n
                  * Group code = 390.\n
                  * \since Introduced in version R2009. */
-        long color_value;
+        int32_t color_value;
                 /*!< A 24-bit color value that should be dealt with in
                  * terms of bytes with values of 0 to 255.\n
                  * The lowest byte is the blue value, the middle byte is
@@ -173,7 +173,7 @@ dxf_acad_proxy_entity_struct
                  * class-level transparency data.\n
                  * Group code = 430.\n
                  * \since Introduced in version R2004. */
-        long transparency;
+        int32_t transparency;
                 /*!< Transparency value.\n
                  * The group code cannot be used by custom entities for
                  * their own data because the group code is reserved for
@@ -275,12 +275,12 @@ int16_t dxf_acad_proxy_entity_get_lineweight (DxfAcadProxyEntity *acad_proxy_ent
 DxfAcadProxyEntity *dxf_acad_proxy_entity_set_lineweight (DxfAcadProxyEntity *acad_proxy_entity, int16_t lineweight);
 char *dxf_acad_proxy_entity_get_plot_style_name (DxfAcadProxyEntity *acad_proxy_entity);
 DxfAcadProxyEntity *dxf_acad_proxy_entity_set_plot_style_name (DxfAcadProxyEntity *acad_proxy_entity, char *plot_style_name);
-long dxf_acad_proxy_entity_get_color_value (DxfAcadProxyEntity *acad_proxy_entity);
-DxfAcadProxyEntity *dxf_acad_proxy_entity_set_color_value (DxfAcadProxyEntity *acad_proxy_entity, long color_value);
+int32_t dxf_acad_proxy_entity_get_color_value (DxfAcadProxyEntity *acad_proxy_entity);
+DxfAcadProxyEntity *dxf_acad_proxy_entity_set_color_value (DxfAcadProxyEntity *acad_proxy_entity, int32_t color_value);
 char *dxf_acad_proxy_entity_get_color_name (DxfAcadProxyEntity *acad_proxy_entity);
 DxfAcadProxyEntity *dxf_acad_proxy_entity_set_color_name (DxfAcadProxyEntity *acad_proxy_entity, char *color_name);
-long dxf_acad_proxy_entity_get_transparency (DxfAcadProxyEntity *acad_proxy_entity);
-DxfAcadProxyEntity *dxf_acad_proxy_entity_set_transparency (DxfAcadProxyEntity *acad_proxy_entity, long transparency);
+int32_t dxf_acad_proxy_entity_get_transparency (DxfAcadProxyEntity *acad_proxy_entity);
+DxfAcadProxyEntity *dxf_acad_proxy_entity_set_transparency (DxfAcadProxyEntity *acad_proxy_entity, int32_t transparency);
 int16_t dxf_acad_proxy_entity_get_original_custom_object_data_format (DxfAcadProxyEntity *acad_proxy_entity);
 DxfAcadProxyEntity *dxf_acad_proxy_entity_set_original_custom_object_data_format (DxfAcadProxyEntity *acad_proxy_entity, int16_t original_custom_object_data_format);
 int32_t dxf_acad_proxy_entity_get_proxy_entity_class_id (DxfAcadProxyEntity *acad_proxy_entity);
